+++ changelog.md
@@ -14,10 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add support for OTA updates of firmware - see OTA_UPDATE.md
 ### Removed
 ### Changed
 ### Fixed
 - Fixed potential buffer overflow error in `bus_send_message()`
+- Always set the `charset=UTF-8` on web responses.
 
 ## [0.0.4] - 2026-01-31